LibPDF: Make all color space classes clamp their inputs
This changes DeviceGrayColorSpace, DeviceRGBColorSpace,
DeviceCMYKColorSpace to clamp their inputs.

LabColorSpace, CalRGBColorSpace, and CalGrayColorSpace did
already clamp -- now they round as well.

The spec only requires this for the latter three as far as I
can tell, but other PDF viewers seem to do this for the first
three as well, so let's follow suit. It also means some other
code in LibPDF that computes colors doesn't have to manually
do this clamping.

For ICCBasedColorSpace, this only adds a FIXME for now.