Kernel/LoopDevice: Avoid creating reference cycles with list nodes
Previously, list nodes would keep the loop device itself alive until the
node was manually removed from the list. This patch breaks that
reference cycle, and allows normal reference counting to take care of
the lifetimes of loop devices (although we do have to leak a reference
on creation because nothing truly owns a loop device).