LibPDF: Change when clips are applied
Previously, before every item we painted, we called activate_clip(),
and after we called deactivate_clip(). These would then temporarily
set a clip rectangle on m_painter.

Now, we update the clip on the painter when a new clipping path is
added (still only the bounding box for now), and also when we
pop a state that had its own clip rect.

This way, we modify the painter's clip much less often than previously.
add_clip_rect() is fast, so this does not make a big difference, but
it will make a big difference once we add support for clipping to
arbitrary paths.

This does feel a bit more brittle, but I can't come up with a way
that does full path clipping at reasonable performance without this
change.

As a side effect, Debug>Apply Clips now also applies to shadings.
That's often confusing, since unclipped shading types 0, 1, and 3
fill the entire page -- but it seems conceptually like the right thing.
Maybe we can tweak this particular thing in some way later.

As another side effect, clips are active the entire time. That makes
it tricky to paint arbitrary information on top of the page. But
that's not something we need to do, so that doesn't seem like a big
problem.

No (intended) effective behavior change.