LibGfx/JBIG2+CCITT: Support decoding MMR-coded JBIG2 graymaps
The main motivation for this is that it's used in page 1 of the test
bitstram in Annex H. Pages 1 and 2 decode to the same bitmap, so
having support for this allows us to compare the two bitmaps, and
we get test coverage for quite a few JBIG2 features that way.

This can also decode amb_2.jb2 from the power jbig2 test dataset.
I'm not aware of any other files using this feature.

An MMR graymap is the only case where the size of the a generic region
is not known in advance, and where the data is immediately followed by
more MMR data. We need to have the MMR decoder skip the EOFB marker at
the end, so that the following bitplanes can be decoded.

To make this possible, we teach the CCITT group 4 decoder to optionally
require the EOFB marker at the end of a stream (to make sure it's
consumed) and let the JBIG2 decoder opt in to that -- but only for
MMR graymaps, since the spec says that the marker is optional in
basically all other cases.

(Since we always pass in number of columns from JBIG2, I think we
shouldn't hit the `has_reached_eol` case from JBIG2.)