Kernel/aarch64: Make RPi mailbox work with arbitrary message addresses
The old code used Memory::virtual_to_low_physical(), which only works
with the initial kernel mappings (including the intial stack when using
pre_init).
It stops working as soon as we reach init_stage2. It also doesn't work
when using the EFI prekernel.

This means that message queue structs no longer need to be aligned
on a 16-byte boundary, as we now copy those into a transfer buffer.