Tests/LibGfx: Set retained bit on symbol segment in JBIG2 symbol tests
This changes the byte at offset 0x30 from 0x00 to 0x01 in all these
files.  That bit stores the segment retention flag of the symbol
segment.  Since that symbol segment is referred to by a text segment
later on, it needs to be retained until then.

(In our decoder, we don't do anything with these segment retention flags
yet. I want to make the loader validate them a bit, which flagged this
being wrong in our test files.)

This only changes files that have only a single symbol segment.