LibMaps: Monitor config even if not monitored in library users
This commit moves the Maps monitor from the Maps application to
the LibMaps library. The File Manager was using LibMaps but it
didn't start the monitor on the Maps domain. So MapsSettings changes
where reflected in the Maps app but not in the File Manager.
This code move results in the all library users always monitor
the config domain correctly.