LibGfx/CCITT: Make EOL always be a valid color change
This is basically the same commit as cefb2d65 but for the encoder. It
seems that you need to consider that the right edge of an image is a
color change for any color.

As mentioned in:
https://github.com/SerenityOS/serenity/pull/23487#discussion_r1515369565

This is not something required for ccitt4.tiff so this is why the test
was passing.