LibGfx: Rename QMCoder to MQCoder
I got the name wrong in #24017.

The history is apparently that Pennebaker, Mitchell, et al introduced
the "Q-coder" in a series of publications at IBM in the late 1980s.

This was then slightly modified and used for arithmetically-coded JPEG
files per itu-t81.pdf (the JPEG spec). The spec doesn't call it that,
but it's referred to as "QM coder" in papers.

JBIG2 and JPEG2000 on the other hand use the "MQ coder", which is
somewhat different from the QM coder (e.g. in how it uses bit stuffing
instead of byte stuffing).

This here is the MQ coder.  The JBIG2 and JPEG2000 specs even call it
that. So rename the file and classes to the right name.

No behavior change.