Kernel: Don't unblock waiters for unregistered processes
If a process isn't completely initialized yet, we should not unblock
waiters for it when it is finalized. No other process (parent or tracee)
should be able to sys$waitid() on it, so we would otherwise have a
leaked zombie child process until the parent is finalized as well.

Similarly, only destroy the WaitBlockerSets for fully initialized
processes, since unregistered process shouldn't be able to create
children.

Currently, all processes have a leaked ref that is balanced by
WaitBlockerSet::finalize(). Since we now only call that function for
registered processes, this ref should only be leaked after
Process:register_new() is called on the process.
The only place where this previously wasn't the case is in sys$fork.