FileManager: Make Launch actions take a strong ref to LauncherHandler
The main functions have a `current_file_handlers` Vector object whose
only job is to sustain the lifetime of LauncherHandlers.
Without it, the ref count would drop to zero after leaving
`add_launch_handler_actions_to_menu()`, and the action would call to a
destroyed object by the time it was activated.

That's because nothing else holds the refcount. The action callback
lambda does capture the objects with `this` keyword, which is a raw
pointer, and not a RefPtr.

So, this commit changes who's in charge of LauncherHandler somewhat.
The action callback now captures a RefPtr to these objects and the
Vectors can go away.