AK: Make `Default` be the default formatting mode for Bytes
There is no way to opt out of the HexDump formatting mode. So let's use
the default `Span` formatter and allow the user to opt in the HexDump
mode for `Span<u8>` aka `Bytes`.

This is mainly useful to print `Vector<u8>`, as the `Vector` formatter
is implemented on top of the `Span` formatter.