Kernel: Panic if trying to block on WaitQueue in critical section or IRQ
Blocking in interrupt handlers or in a critical section is never a good
idea since it can easily lead to deadlocks.

Additionally, maybe_block() won't work properly in such situations, as
Scheduler::yield() will delay the context switch until the interrupt
handler or critical section is done, breaking the assumption that this
function will block if the thread state is set to Blocked.