Tests/LibGfx: Convert bitmap-symbol.jbig2 to a json-based test
I tried to make all the numbers (both explicit and implicit) match the
ones I used back in #23659.

I also put in all flags, huffman_flags, etc, even though most of the
values are set to their default values and don't have to be specified.
I checked by building `image` with `JBIG2_DEBUG 1` set and added a few
additional local dbgln()s and compared the output on it reading both old
and new versions.

The generated jbig2 file is fairly similar to the previous one, but it's
missing a pointless generic region that `jbig2` had put in.

Also, Preview.app can now decode the file (when run through
Meta/jbig2_to_pdf.py), which it couldn't do before. This is probably
because the arithmetic coded data now ends in "ffac" instead of just
"ff".

As far as I understand, this file is strictly speaking invalid, as I
think all symbol instances in a strip are supposed to be at most
strip_size apart. I might update it in the future. (But all decoders
seem to decode it fine.)