Tests/LibGfx: Jsonify JBIG2 text refinement test with delta width < 0
The new test is identical to bitmap-symbol-textrefine.json, except that
the refinement on the instance has delta_width of -1, and the symbol
it's refining is 1 wider (width of 41 instead of 40).

Remove the old, non-json-based test from the first commit in #25751.
I verified that tests still fail if I locally revert the second commit
in #25751.