LibGfx: Make Bayer dithering more luminosity-conservative
On average, this patch reduces the error in luminosity between an input
image and its dithered equivalent. This is done by correcting an off-by-
one error in the threshold computation and also by making sure the range
of mapped gray values to matrix pattern is of equal size for all ranges.
This means that all values between 0 and 255 / (N * N + 1) (51 for
Bayer2x2) will result in a black pixel while only a value of 0 would
previously do it.

The test works by first generating gray bitmap, then applying Bayer
dithering on them and finally compare the luminosity of the two.